+++ packages/authhero/test/routes/management-api/connections.spec.ts
@@ -14,26 +14,25 @@ describe("connections", () => {
     // --------------------------------------------
     // POST
     // --------------------------------------------
-    const createConnectionResponse =
-      await managementClient.api.v2.connections.$post(
-        {
-          json: {
-            name: "apple",
-            strategy: "apple",
-            options: {
-              team_id: "teamId",
-            },
-          },
-          header: {
-            "tenant-id": "tenantId",
+    const createConnectionResponse = await managementClient.connections.$post(
+      {
+        json: {
+          name: "apple",
+          strategy: "apple",
+          options: {
+            team_id: "teamId",
           },
         },
-        {
-          headers: {
-            authorization: `Bearer ${token}`,
-          },
+        header: {
+          "tenant-id": "tenantId",
         },
-      );
+      },
+      {
+        headers: {
+          authorization: `Bearer ${token}`,
+        },
+      },
+    );
 
     expect(createConnectionResponse.status).toBe(201);
     const createdConnection = await createConnectionResponse.json();
@@ -54,7 +53,7 @@ describe("connections", () => {
     // --------------------------------------------
     // PATCH
     // --------------------------------------------
-    const updateConnectionResponse = await managementClient.api.v2.connections[
+    const updateConnectionResponse = await managementClient.connections[
       ":id"
     ].$patch(
       {
@@ -78,8 +77,31 @@ describe("connections", () => {
     );
 
     expect(updateConnectionResponse.status).toBe(200);
-    const updatedConnection =
+    const updateConnection =
       (await updateConnectionResponse.json()) as Connection;
+    expect(updateConnection.options).toEqual({
+      team_id: "teamId2",
+    });
+
+    const updatesConnectionResponse = await managementClient.connections[
+      ":id"
+    ].$get(
+      {
+        param: {
+          id: id!,
+        },
+        header: {
+          "tenant-id": "tenantId",
+        },
+      },
+      {
+        headers: {
+          authorization: `Bearer ${token}`,
+        },
+      },
+    );
+    expect(updatesConnectionResponse.status).toBe(200);
+    const updatedConnection = await updatesConnectionResponse.json();
     expect(updatedConnection.options).toEqual({
       team_id: "teamId2",
     });
@@ -87,7 +109,7 @@ describe("connections", () => {
     // --------------------------------------------
     // DELETE
     // --------------------------------------------
-    const deleteConnectionResponse = await managementClient.api.v2.connections[
+    const deleteConnectionResponse = await managementClient.connections[
       ":id"
     ].$delete(
       {
@@ -110,20 +132,19 @@ describe("connections", () => {
     // --------------------------------------------
     // LIST
     // --------------------------------------------
-    const listConnectionsResponse =
-      await managementClient.api.v2.connections.$get(
-        {
-          query: {},
-          header: {
-            "tenant-id": "tenantId",
-          },
+    const listConnectionsResponse = await managementClient.connections.$get(
+      {
+        query: {},
+        header: {
+          "tenant-id": "tenantId",
         },
-        {
-          headers: {
-            authorization: `Bearer ${token}`,
-          },
+      },
+      {
+        headers: {
+          authorization: `Bearer ${token}`,
         },
-      );
+      },
+    );
 
     expect(listConnectionsResponse.status).toBe(200);
     const connections = await listConnectionsResponse.json();