+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @authhero/adapter-interfaces
 
+## 0.25.0
+
+### Minor Changes
+
+- added email providers and removed tickets
+- Added email providers
+
+## 0.24.0
+
+### Minor Changes
+
+- add code verifier to codes table
+
+## 0.23.0
+
+### Minor Changes
+
+- make strategy mandatory for connections
+
 ## 0.22.1
 
 ### Patch Changes